Refactor sort dir structure and remove genmap-gslib.h
