+++ src/genmap-types.h
@@ -2,7 +2,7 @@
 #define _GENMAP_TYPES_H_
 
 #include <float.h>
-#include <genmap-gslib.h>
+#include <parRSB.h>
 
 typedef long long GenmapLong;
 typedef unsigned long long GenmapULong;
