+++ src/sort/sort-impl.h
@@ -9,11 +9,33 @@
 #define min(a, b) ((a) < (b) ? (a) : (b))
 #define max(a, b) ((a) > (b) ? (a) : (b))
 
+/*
+ * local sort
+ */
+
+int sort_local(struct sort *s);
+
+/*
+ * parallel sort
+ */
+
+struct hypercube {
+  struct sort *data;
+  int nprobes;
+  double *probes;
+  ulong *probe_cnt;
+};
+
+int parallel_bin_sort(struct sort *s, struct comm *c);
+int parallel_hypercube_sort(struct hypercube *h, struct comm *c);
+
+/*
+ * Auxiliary functions
+ */
 double get_scalar(struct array *a, uint i, uint off, uint usize, gs_dom type);
 void get_extrema(void *extrema, struct sort *s, uint field, struct comm *c);
+
 int set_dest(uint *proc, uint np, ulong start, uint size, ulong nelem);
 int load_balance(struct array *a, size_t size, struct comm *c,
                  struct crystal *cr);
-int sort_local(struct sort *s);
-
 #endif
